feat: added base processor
