+++ backend/src/processors/extract_processor.py
@@ -1,17 +1,14 @@
 from dataclasses import dataclass
 from typing import List, Dict, Optional
 import json
-import logging
-from datetime import datetime
 import os
-from langchain_aws.chat_models import ChatBedrock
 from langchain.prompts import ChatPromptTemplate
 from langchain_core.messages import HumanMessage, SystemMessage
 import concurrent.futures
 
 from ..models.models import ParseVersion, ExtractVersion
-from ..llm.llm_factory import LLMConfig, LLMFactory
-from ..core.config import ConfigManager, FILE_SYSTEM
+from ..core.config import ConfigManager
+from .base_processor import BaseProcessor, ProcessingResult
 
 
 @dataclass
@@ -26,15 +23,7 @@ class Entity:
     related_concepts: List[str]
 
 
-@dataclass
-class ProcessingResult:
-    success: bool
-    status: str
-    message: str
-    error: str
-
-
-class ExtractProcessor:
+class ExtractProcessor(BaseProcessor):
     def __init__(
         self,
         parse_version: ParseVersion,
@@ -53,78 +42,7 @@ class ExtractProcessor:
         self.entity_details_prompt = self.extract_version.entity_details_prompt
         # TODO: Implement custom instructions
         self.custom_instructions = self.extract_version.custom_instructions
-        self.config = config_manager
-        self._setup_directories()
-        self.logger = self._setup_logger()
-        self.model = self._setup_model()
-
-    def _setup_model(self) -> ChatBedrock:
-        """Initialize the LLM model"""
-        llm_config = LLMConfig(
-            provider=self.config.get("llm_provider", "bedrock"),
-            profile_name=self.config.get("aws_profile"),
-            model_id=self.config.get(
-                "aws_model_id", "us.anthropic.claude-3-5-sonnet-20241022-v2:0"
-            ),
-            model_kwargs={
-                "temperature": float(self.config.get("llm_temperature", 0)),
-                "max_tokens": int(self.config.get("llm_max_tokens", 4096)),
-            },
-        )
-        return LLMFactory(llm_config).create_model()
-
-    def _setup_logger(self) -> logging.Logger:
-        """Setup logging with proper error handling."""
-        logger = logging.getLogger(
-            f"EntityProcessor_{self.extract_version.pipeline_id}"
-        )
-        logger.setLevel(logging.DEBUG)
-
-        # Remove existing handlers
-        for handler in logger.handlers[:]:
-            logger.removeHandler(handler)
-
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        log_file = os.path.join(
-            self.output_dir, "logs", f"extract_processor_{timestamp}.log"
-        )
-        formatter = logging.Formatter(
-            "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-        )
-
-        try:
-            file_handler = logging.FileHandler(log_file)
-            file_handler.setLevel(logging.DEBUG)
-            file_handler.setFormatter(formatter)
-            logger.addHandler(file_handler)
-
-            console_handler = logging.StreamHandler()
-            console_handler.setLevel(logging.INFO)
-            console_handler.setFormatter(formatter)
-            logger.addHandler(console_handler)
-
-            logger.info(f"Logging initialized. Log file: {log_file}")
-        except Exception as e:
-            console_handler = logging.StreamHandler()
-            console_handler.setLevel(logging.DEBUG)
-            console_handler.setFormatter(formatter)
-            logger.addHandler(console_handler)
-            logger.error(
-                f"Failed to create file handler: {str(e)}. Using console logging only."
-            )
-
-        return logger
-
-    def _setup_directories(self):
-        """Create all necessary directories for processing."""
-        if self.config.get("file_system", "local") == "local":
-            directories = [
-                self.output_dir,
-                os.path.join(self.output_dir, "logs"),
-                os.path.join(self.output_dir, "temp"),
-            ]
-            for directory in directories:
-                os.makedirs(directory, exist_ok=True)
+        super().__init__(config_manager)
 
     def _initial_extraction(self, content: str) -> tuple[List[str], List[Dict]]:
         """Perform initial extraction of entities and relationships."""
