+++ backend/src/processors/merge_processor.py
@@ -1,26 +1,16 @@
-from dataclasses import dataclass
-from typing import List, Dict, Optional
+from typing import List, Dict
 import json
 import os
 import logging
 from datetime import datetime
-from langchain_aws.chat_models import ChatBedrock
 from langchain.prompts import ChatPromptTemplate
 from langchain_core.messages import HumanMessage, SystemMessage
 from ..models.models import ExtractVersion, MergeVersion
-from ..llm.llm_factory import LLMConfig, LLMFactory
-from ..core.config import FILE_SYSTEM, ConfigManager
+from ..core.config import ConfigManager
+from .base_processor import BaseProcessor, ProcessingResult
 
 
-@dataclass
-class ProcessingResult:
-    success: bool
-    status: str
-    message: str
-    error: str
-
-
-class MergeProcessor:
+class MergeProcessor(BaseProcessor):
     def __init__(
         self,
         extract_versions: List[ExtractVersion],
@@ -34,72 +24,7 @@ class MergeProcessor:
         self.entity_merge_prompt = self.merge_version.entity_merge_prompt
         # TODO: Implement custom instructions
         self.custom_instructions = self.merge_version.custom_instructions
-        self.config = config_manager
-        self._setup_directories()
-        self.logger = self._setup_logger()
-        self.model = self._setup_model()
-
-    def _setup_model(self) -> ChatBedrock:
-        """Initialize the LLM model based on domain configuration"""
-        llm_config = LLMConfig(
-            provider=self.config.get("llm_provider", "bedrock"),
-            profile_name=self.config.get("aws_profile"),
-            model_id=self.config.get(
-                "aws_model_id", "us.anthropic.claude-3-5-sonnet-20241022-v2:0"
-            ),
-            model_kwargs={
-                "temperature": float(self.config.get("llm_temperature", 0)),
-                "max_tokens": int(self.config.get("llm_max_tokens", 4096)),
-            },
-        )
-        return LLMFactory(llm_config).create_model()
-
-    def _setup_logger(self) -> logging.Logger:
-        logger = logging.getLogger(f"MergeProcessor_{self.merge_version.pipeline_id}")
-        logger.setLevel(logging.DEBUG)
-
-        for handler in logger.handlers[:]:
-            logger.removeHandler(handler)
-
-        formatter = logging.Formatter(
-            "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-        )
-        log_file = os.path.join(
-            self.output_dir,
-            "logs",
-            f"merge_processor_{datetime.now().strftime('%Y%m%d_%H%M%S')}.log",
-        )
-
-        try:
-            file_handler = logging.FileHandler(log_file)
-            file_handler.setLevel(logging.DEBUG)
-            file_handler.setFormatter(formatter)
-            logger.addHandler(file_handler)
-
-            console_handler = logging.StreamHandler()
-            console_handler.setLevel(logging.INFO)
-            console_handler.setFormatter(formatter)
-            logger.addHandler(console_handler)
-
-            logger.info(f"Logging initialized: {log_file}")
-        except Exception as e:
-            console_handler = logging.StreamHandler()
-            console_handler.setLevel(logging.DEBUG)
-            console_handler.setFormatter(formatter)
-            logger.addHandler(console_handler)
-            logger.error(f"Failed to create file handler: {str(e)}")
-
-        return logger
-
-    def _setup_directories(self):
-        if self.config.get("file_system", "local") == "local":
-            directories = [
-                self.output_dir,
-                os.path.join(self.output_dir, "logs"),
-                os.path.join(self.output_dir, "temp"),
-            ]
-            for directory in directories:
-                os.makedirs(directory, exist_ok=True)
+        super().__init__(config_manager)
 
     def _merge_batch(self, entities_batch: List[Dict]) -> Dict:
         content = [
