+++ backend/src/processors/parse_processor.py
@@ -3,13 +3,9 @@ from PIL import Image
 import base64
 from io import BytesIO
 import os
-from dataclasses import dataclass
 from typing import List, Dict
 import json
-import logging
-from datetime import datetime
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from langchain_aws.chat_models import ChatBedrock
 from langchain.prompts import ChatPromptTemplate
 from langchain_core.messages import HumanMessage, SystemMessage
 
@@ -17,20 +13,11 @@ from ..models.models import (
     ParseVersion,
     FileVersion,
 )
-from ..llm.llm_factory import LLMConfig, LLMFactory
-from ..core.config import FILE_SYSTEM, ConfigManager
+from ..core.config import ConfigManager
+from .base_processor import BaseProcessor, ProcessingResult
 
 
-# TODO: Implement quality score and improvement message
-@dataclass
-class ProcessingResult:
-    success: bool
-    status: str
-    message: str
-    error: str
-
-
-class ParseProcessor:
+class ParseProcessor(BaseProcessor):
     def __init__(
         self,
         file_version: FileVersion,
@@ -44,87 +31,13 @@ class ParseProcessor:
         self.convert_prompt = self.parse_version.convert_prompt
         # TODO: Implement custom instructions
         self.custom_instructions = self.parse_version.custom_instructions
-        self.config = config_manager
-        self._setup_directories()
-        self.logger = self._setup_logger()
-        self.model = self._setup_model()
+        super().__init__(config_manager)
 
-    def _setup_model(self) -> ChatBedrock:
-        """Initialize the LLM model"""
-        llm_config = LLMConfig(
-            provider=self.config.get("llm_provider", "bedrock"),
-            profile_name=self.config.get("aws_profile"),
-            model_id=self.config.get(
-                "aws_model_id", "us.anthropic.claude-3-5-sonnet-20241022-v2:0"
-            ),
-            model_kwargs={
-                "temperature": float(self.config.get("llm_temperature", 0)),
-                "max_tokens": int(self.config.get("llm_max_tokens", 4096)),
-            },
-        )
-        return LLMFactory(llm_config).create_model()
+    def _get_output_dir(self) -> str:
+        return self.parse_version.output_dir
 
-    def _setup_logger(self) -> logging.Logger:
-        """Setup logging for the processor with proper directory handling."""
-        logger = logging.getLogger(
-            f"ParseProcessor_{self.file_version.file_version_id}"
-        )
-        logger.setLevel(logging.DEBUG)
-
-        # Remove any existing handlers to avoid duplicate logging
-        for handler in logger.handlers[:]:
-            logger.removeHandler(handler)
-
-        # Create logs directory path
-        logs_dir = os.path.join(self.parse_version.output_dir, "logs")
-
-        # File handler
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        log_file = os.path.join(logs_dir, f"parse_processor_{timestamp}.log")
-
-        try:
-            file_handler = logging.FileHandler(log_file)
-            file_handler.setLevel(logging.DEBUG)
-            formatter = logging.Formatter(
-                "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-            )
-            file_handler.setFormatter(formatter)
-            logger.addHandler(file_handler)
-
-            # Add a stream handler for console output
-            console_handler = logging.StreamHandler()
-            console_handler.setLevel(logging.INFO)
-            console_handler.setFormatter(formatter)
-            logger.addHandler(console_handler)
-
-            logger.info(f"Logging initialized. Log file: {log_file}")
-        except Exception as e:
-            # Fallback to console-only logging if file handler fails
-            console_handler = logging.StreamHandler()
-            console_handler.setLevel(logging.DEBUG)
-            console_handler.setFormatter(formatter)
-            logger.addHandler(console_handler)
-            logger.error(
-                f"Failed to create file handler: {str(e)}. Falling back to console logging only."
-            )
-
-        return logger
-
-    def _setup_directories(self):
-        """Create all necessary directories for processing and logging."""
-        if self.config.get("file_system", "local") == "local":
-            # Create base output directory
-            os.makedirs(self.parse_version.output_dir, exist_ok=True)
-
-            # Create subdirectories
-            directories = [
-                os.path.join(self.parse_version.output_dir, "images"),
-                os.path.join(self.parse_version.output_dir, "logs"),
-                os.path.join(self.parse_version.output_dir, "temp"),
-            ]
-
-            for directory in directories:
-                os.makedirs(directory, exist_ok=True)
+    def _get_logger_name(self) -> str:
+        return f"ParseProcessor_{self.file_version.file_version_id}"
 
     def _convert_page_to_image(self, page: fitz.Page) -> Image.Image:
         """Convert PDF page to PIL Image."""
